feat(schemas): add UnicodeVersionSchema and refactor imports
- Introduced `UnicodeVersionSchema` to define the structure of Unicode version data.
- Refactored imports in `index.ts` to explicitly export `UnicodeVersionSchema`.
- Removed the old schema definition from `schemas.ts` and replaced it with a reference to the new schema.
