+++ packages/schemas/src/index.ts
@@ -1,1 +1,5 @@
-export * from "./fs";
+export { FileEntrySchema, UCDStoreManifestSchema } from "./fs";
+export type { FileEntry, UCDStoreManifest } from "./fs";
+
+export { UnicodeVersionSchema } from "./unicode";
+export type { UnicodeVersion } from "./unicode";
